frameworks/native: surfaceflinger: Do not wait for vsync for MDP Composition
SurfaceFlinger waits for next vsync cycle if a frame
is available from client. For MDP 3.0 targets, we wait
for vsync cycle in every composition step. If a frame
if available from client just after vsync, we would delay
it showing on display by one vsync cycle also making client
wait in DQ. As MDP composition waits for vsync, remove wait
for vsync in composition step when in MDP Composition.

Change-Id: If0054e8efc9fcfb8c621f45f130e054a799703c4